+++ src/main/java/io/github/battlepass/objects/user/QuestStore.java
@@ -1,12 +1,13 @@
 package io.github.battlepass.objects.user;
 
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class QuestStore {
-    private final ConcurrentHashMap<String, ConcurrentHashMap<String, Integer>> quests = new ConcurrentHashMap<>();
+    private final ConcurrentHashMap<String, ConcurrentHashMap<String, BigInteger>> quests = new ConcurrentHashMap<>();
 
-    public Map<String, ConcurrentHashMap<String, Integer>> asMap() {
+    public Map<String, ConcurrentHashMap<String, BigInteger>> asMap() {
         return this.quests;
     }
